Use built-in implementations of Widget.Children
