+++ redwood-treehouse/src/iosMain/kotlin/app/cash/treehouse/IosTreehouseDispatchers.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright (C) 2022 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.treehouse
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import platform.Foundation.NSThread
+
+// TODO(jwilson): we're currently doing everything on the main thread on iOS.
+public class IosTreehouseDispatchers : TreehouseDispatchers {
+  override val main: CoroutineDispatcher = Dispatchers.Main
+  override val zipline: CoroutineDispatcher = Dispatchers.Main
+
+  override fun checkMain() {
+    check(NSThread.isMainThread)
+  }
+
+  override fun checkZipline() {
+    checkMain()
+  }
+}
